Refactor useApp
