+++ src/components/forms/pool_actions/WithdrawForm/components/ProportionalWithdrawalInputV2.vue
@@ -47,6 +47,7 @@ const {
   exitTokenInfo,
   fiatAmountsOut,
   fiatTotalOut,
+  transactionDeadline,
 } = useExitPool();
 const { t } = useI18n();
